+++ documentation/docs/changelog.md
@@ -4,6 +4,7 @@
 
 * Add the CommandsHandler class to allow addons to register commands
 * Add the Output class to replace print() calls
+* Add the RaidMarker model class and the library's possible marker instances
 * Allow passing addon properties to the library
 
 ## 2024.03.16 - version 0.0.6-alpha