+++ documentation/docs/changelog.md
@@ -4,6 +4,7 @@
 
 * Add the Events facade class to serve as a way to improve event handling
     * Listen to `PLAYER_LOGIN`
+    * Listen to `PLAYER_TARGET_CHANGED` and split it into three events
 * Add the CommandsHandler class to allow addons to register commands
 * Add the Output class to replace print() calls
 * Add the RaidMarker model class and the library's possible marker instances